+++ assets/translations/en-us.yaml
@@ -2148,6 +2148,7 @@ rbac:
     fleetworkspace-member: Member
     fleetworkspace-readonly: Read-Only
   roletemplate:
+    label: Roles
     newUserDefault:
       no: No
       tooltip: This does not affect any bindings to the role that already exist.
@@ -2165,18 +2166,18 @@ rbac:
           apiGroups: API Groups
     subtypes:
       GLOBAL:
+        createButton: Create Global Role
         label: Global
-        abbreviation: G
         yes: "Yes: Default role for new users"
         defaultLabel: New User Default
       CLUSTER:
+        createButton: Create Cluster Role
         label: Cluster
-        abbreviation: C
         yes: "Yes: Default role for new cluster creation"
         defaultLabel: Cluster Creator Default
       NAMESPACE:
+        createButton: Create Project/Namespace Role
         label: Project/Namespace
-        abbreviation: PN
         yes: "Yes: Default role for new project creation"
         defaultLabel: Project Creator Default
       noContext:
@@ -2284,6 +2285,7 @@ resourceDetail:
       {annotations, plural,
       =1 {Show 1 annotation}
       other {Show {annotations} annotations}}
+    name: Name
   header:
     clone: "Clone from {subtype} {name}"
     create: Create {subtype}
@@ -2875,6 +2877,7 @@ tableHeaders:
   message: Message
   minKubernetesVersion: Min Kubernetes Version
   name: Name
+  nameDisplay: Display Name
   nameUnlinked: Name
   namespace: Namespace
   namespaceName: Name
