+++ detail/management.cattle.io.globalrole.vue
@@ -1,0 +1,8 @@
+<script>
+import RoleDetailEdit from '@/components/auth/RoleDetailEdit';
+
+export default { components: { RoleDetailEdit } };
+</script>
+<template>
+  <RoleDetailEdit v-bind="$attrs" />
+</template>
