+++ pages/c/_cluster/auth/roles/index.vue
@@ -1,0 +1,172 @@
+<script>
+import Tab from '@/components/Tabbed/Tab';
+import Tabbed from '@/components/Tabbed';
+import { MANAGEMENT } from '@/config/types';
+import ResourceTable from '@/components/ResourceTable';
+import Loading from '@/components/Loading';
+import { SUBTYPE_MAPPING } from '@/models/management.cattle.io.roletemplate';
+import { NAME } from '@/config/product/auth';
+
+const GLOBAL = SUBTYPE_MAPPING.GLOBAL.key;
+const CLUSTER = SUBTYPE_MAPPING.CLUSTER.key;
+const PROJECT = SUBTYPE_MAPPING.NAMESPACE.key;
+
+const createGlobalRole = {
+  name:   'c-cluster-product-resource-create',
+  params: {
+    cluster:  'local',
+    product:  NAME,
+    resource: MANAGEMENT.GLOBAL_ROLE,
+  }
+};
+const createRoleTemplate = {
+  name:   'c-cluster-product-resource-create',
+  params: {
+    cluster:  'local',
+    product:  NAME,
+    resource: MANAGEMENT.ROLE_TEMPLATE,
+  }
+};
+
+export default {
+  components: {
+    Tab, Tabbed, ResourceTable, Loading
+  },
+
+  async asyncData({ store }) {
+    const globalRoleSchema = store.getters[`management/schemaFor`](MANAGEMENT.GLOBAL_ROLE);
+    const roleTemplatesSchema = store.getters[`management/schemaFor`](MANAGEMENT.ROLE_TEMPLATE);
+
+    return {
+      globalRoles:   globalRoleSchema ? await store.dispatch(`management/findAll`, { type: MANAGEMENT.GLOBAL_ROLE }) : [],
+      roleTemplates: roleTemplatesSchema ? await store.dispatch(`management/findAll`, { type: MANAGEMENT.ROLE_TEMPLATE }) : [],
+    };
+  },
+
+  data() {
+    const globalRoleSchema = this.$store.getters[`management/schemaFor`](MANAGEMENT.GLOBAL_ROLE);
+    const roleTemplatesSchema = this.$store.getters[`management/schemaFor`](MANAGEMENT.ROLE_TEMPLATE);
+
+    return {
+      tabs: {
+        [GLOBAL]: {
+          canFetch:       globalRoleSchema?.collectionMethods.find(verb => verb === 'GET'),
+          canCreate:      globalRoleSchema?.resourceMethods.find(verb => verb === 'PUT'),
+          weight:         3,
+          labelKey:       SUBTYPE_MAPPING.GLOBAL.labelKey,
+          schema:         globalRoleSchema,
+          createLocation: {
+            ...createGlobalRole,
+            query: { roleContext: GLOBAL }
+          },
+        },
+        [CLUSTER]: {
+          canFetch:       roleTemplatesSchema?.collectionMethods.find(verb => verb === 'GET'),
+          canCreate:      roleTemplatesSchema?.resourceMethods.find(verb => verb === 'PUT'),
+          labelKey:       SUBTYPE_MAPPING.CLUSTER.labelKey,
+          weight:         2,
+          schema:         roleTemplatesSchema,
+          createLocation: {
+            ...createRoleTemplate,
+            query: { roleContext: CLUSTER }
+          },
+        },
+        [PROJECT]: {
+          canFetch:       roleTemplatesSchema?.collectionMethods.find(verb => verb === 'GET'),
+          canCreate:      roleTemplatesSchema?.resourceMethods.find(verb => verb === 'PUT'),
+          labelKey:       SUBTYPE_MAPPING.NAMESPACE.labelKey,
+          weight:         1,
+          schema:         roleTemplatesSchema,
+          createLocation: {
+            ...createRoleTemplate,
+            query: { roleContext: PROJECT }
+          },
+        },
+      },
+
+      GLOBAL,
+      CLUSTER,
+      PROJECT,
+
+      globalRoles:   null,
+      roleTemplates: null,
+    };
+  },
+
+  computed: {
+    globalResources() {
+      return this.globalRoles;
+    },
+
+    clusterResources() {
+      return this.roleTemplates.filter(r => r.context === SUBTYPE_MAPPING.CLUSTER.context);
+    },
+
+    namespaceResources() {
+      return this.roleTemplates.filter(r => r.context === SUBTYPE_MAPPING.NAMESPACE.context);
+    },
+
+    type() {
+      if (this.$route.hash.endsWith(PROJECT)) {
+        return PROJECT;
+      }
+      if (this.$route.hash.endsWith(CLUSTER)) {
+        return CLUSTER;
+      }
+
+      return GLOBAL;
+    },
+
+    canCreate() {
+      return this.tabs[this.type].canCreate;
+    },
+
+    createLabel() {
+      return this.t(`rbac.roletemplate.subtypes.${ this.type }.createButton`);
+    },
+
+    createLocation() {
+      return this.tabs[this.type].createLocation;
+    }
+
+  },
+
+};
+</script>
+
+<template>
+  <Loading v-if="!globalRoles || !roleTemplates" />
+  <div v-else>
+    <header>
+      <div class="title">
+        <h1 class="m-0">
+          {{ t('rbac.roletemplate.label') }}
+        </h1>
+      </div>
+      <div class="actions-container">
+        <div class="actions">
+          <n-link
+            v-if="canCreate"
+            :to="createLocation"
+            class="btn role-primary"
+          >
+            {{ createLabel }}
+          </n-link>
+        </div>
+      </div>
+    </header>
+    <Tabbed>
+      <Tab v-if="tabs[GLOBAL].canFetch" :name="GLOBAL" :weight="tabs[GLOBAL].weight" :label-key="tabs[GLOBAL].labelKey">
+        <ResourceTable :schema="tabs[GLOBAL].schema" :rows="globalResources" />
+      </Tab>
+
+      <Tab v-if="tabs[CLUSTER].canFetch" :name="CLUSTER" :weight="tabs[CLUSTER].weight" :label-key="tabs[CLUSTER].labelKey">
+        <ResourceTable :schema="tabs[CLUSTER].schema" :rows="clusterResources" />
+      </Tab>
+
+      <Tab v-if="tabs[PROJECT].canFetch" :name="PROJECT" :weight="tabs[PROJECT].weight" :label-key="tabs[PROJECT].labelKey">
+        <ResourceTable :schema="tabs[PROJECT].schema" :rows="namespaceResources" />
+      </Tab>
+    </Tabbed>
+  </div>
+</template>
